chore(test): fix flaky query cache flush tests (#2035)
